MS-205 Move code related to sync from id module to sync
