+++ .github/workflows/pr-checks.yml
@@ -14,7 +14,6 @@ jobs:
         secrets: inherit
         with:
             modules: |
-                id
                 infra:core
                 infra:ui-base
                 infra:network
