+++ id/src/main/java/com/simprints/id/Application.kt
@@ -11,6 +11,7 @@ import com.simprints.core.tools.utils.LanguageHelper
 import com.simprints.infra.logging.LoggingConstants.CrashReportingCustomKeys.DEVICE_ID
 import com.simprints.infra.logging.Simber
 import com.simprints.infra.logging.SimberBuilder
+import com.simprints.infra.sync.SyncOrchestrator
 import dagger.hilt.android.HiltAndroidApp
 import io.reactivex.exceptions.UndeliverableException
 import io.reactivex.plugins.RxJavaPlugins
@@ -26,10 +27,7 @@ open class Application : CoreApplication(), Configuration.Provider {
     lateinit var workerFactory: HiltWorkerFactory
 
     @Inject
-    lateinit var cleanupDeprecatedWorkers: CleanupDeprecatedWorkersUseCase
-
-    @Inject
-    lateinit var scheduleBackgroundSync: ScheduleBackgroundSyncUseCase
+    lateinit var syncOrchestrator: SyncOrchestrator
 
     @AppScope
     @Inject
@@ -57,8 +55,8 @@ open class Application : CoreApplication(), Configuration.Provider {
         SimberBuilder.initialize(this)
         Simber.tag(DEVICE_ID, true).i(deviceHardwareId)
         appScope.launch {
-            cleanupDeprecatedWorkers()
-            scheduleBackgroundSync()
+            syncOrchestrator.cleanupWorkers()
+            syncOrchestrator.scheduleBackgroundWork()
         }
     }
