+++ infra/sync/src/main/java/com/simprints/infra/sync/SyncOrchestrator.kt
@@ -7,4 +7,5 @@ interface SyncOrchestrator {
 
     fun startDeviceSync()
 
+    fun cleanupWorkers()
 }
