+++ infra/sync/src/main/java/com/simprints/infra/sync/SyncOrchestratorImpl.kt
@@ -6,11 +6,13 @@ import com.simprints.infra.sync.extensions.schedulePeriodicWorker
 import com.simprints.infra.sync.extensions.startWorker
 import com.simprints.infra.sync.config.worker.DeviceConfigDownSyncWorker
 import com.simprints.infra.sync.config.worker.ProjectConfigDownSyncWorker
+import com.simprints.infra.sync.usecase.CleanupDeprecatedWorkersUseCase
 import javax.inject.Inject
 
 internal class SyncOrchestratorImpl @Inject constructor(
     private val workManager: WorkManager,
     private val authStore: AuthStore,
+    private val cleanupDeprecatedWorkers: CleanupDeprecatedWorkersUseCase,
 ) : SyncOrchestrator {
 
     override suspend fun scheduleBackgroundWork() {
@@ -23,6 +25,9 @@ internal class SyncOrchestratorImpl @Inject constructor(
                 SyncConstants.DEVICE_SYNC_WORK_NAME,
                 SyncConstants.DEVICE_SYNC_REPEAT_INTERVAL
             )
+            // TODO eventSyncManager.scheduleSync()
+            // TODO imageUpSyncScheduler.scheduleImageUpSync()
+            // TODO firmwareFileUpdateScheduler.scheduleOrCancelWorkIfNecessary()
         }
     }
 
@@ -33,5 +38,10 @@ internal class SyncOrchestratorImpl @Inject constructor(
 
     override fun startDeviceSync() {
         workManager.startWorker<DeviceConfigDownSyncWorker>(SyncConstants.DEVICE_SYNC_WORK_NAME_ONE_TIME)
+    }
+
+
+    override fun cleanupWorkers() {
+        cleanupDeprecatedWorkers()
     }
 }
